reverse weird structural decisions
